+++ src/sculptor/__init__.py
@@ -10,8 +10,10 @@ __version__ = "0.1.0"
 # Optionally, import top-level classes/functions
 from .sculptor import Sculptor
 from .sculptor_pipeline import SculptorPipeline
+from .async_sculptor import AsyncSculptor
 
 __all__ = [
     "Sculptor",
-    "SculptorPipeline"
+    "SculptorPipeline",
+    "AsyncSculptor"
 ]
